Moved the adapter and adapter folder to active_model_serializers folder and changed the module namespace
