--- lib/active_model/serializer/adapter/json/fragment_cache.rb
@@ -1,13 +1,0 @@
-module ActiveModel
-  class Serializer
-    module Adapter
-      class Json
-        class FragmentCache
-          def fragment_cache(cached_hash, non_cached_hash)
-            non_cached_hash.merge cached_hash
-          end
-        end
-      end
-    end
-  end
-end
