--- lib/active_model/serializer/adapter/null.rb
@@ -1,11 +1,0 @@
-module ActiveModel
-  class Serializer
-    module Adapter
-      class Null < Base
-        def serializable_hash(options = nil)
-          {}
-        end
-      end
-    end
-  end
-end
