+++ lib/active_model_serializers/adapter/json/fragment_cache.rb
@@ -1,0 +1,11 @@
+module ActiveModelSerializers
+  module Adapter
+    class Json
+      class FragmentCache
+        def fragment_cache(cached_hash, non_cached_hash)
+          non_cached_hash.merge cached_hash
+        end
+      end
+    end
+  end
+end
