+++ lib/active_model_serializers/adapter/null.rb
@@ -1,0 +1,9 @@
+module ActiveModelSerializers
+  module Adapter
+    class Null < Base
+      def serializable_hash(options = nil)
+        {}
+      end
+    end
+  end
+end
